fix(model policies): Support float weights (#1163)

+++ src/ibm_watsonx_orchestrate/agent_builder/model_policies/types.py
@@ -17,7 +17,7 @@ class ModelPolicyRetry(BaseModel):
     on_status_codes: List[int] = None
 
 class ModelPolicyTarget(BaseModel):
-    weight: int = None
+    weight: float = None
     model_name: str = None
 
 class ModelPolicyInner(BaseModel):
